added server boileplate. Created heroku app in directory. reorganized react app (client)
