+++ client/src/containers/App.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <h1>Hello. I will someday become a video game trader</h1>
+    </div>
+  );
+}
+
+export default App;
